requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.7.8
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
dataclasses-json==0.6.3
distlib==0.3.8
distro==1.8.0
dulwich==0.21.7
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.12.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
installer==0.7.0
jaraco.classes==3.3.0
jsonpatch==1.33
jsonpointer==2.4
-e git+https://github.com/QAInsights/kel.git@083289ed928ee9cabc7e7a6899f0878f089fca3f#egg=kel_cli
keyring==24.3.0
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.0
langsmith==0.0.70
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.6
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
proto-plus==1.23.0
protobuf==4.25.1
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pyperclip==1.8.2
pyproject_hooks==1.0.0
PyYAML==6.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
shellingham==1.5.4
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.1
trove-classifiers==2023.11.29
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
xattr==0.10.1
yarl==1.9.4
zipp==3.17.0
